add test

+++ test/yosupo/sort_points_by_argument.test.cpp
@@ -14,7 +14,7 @@ signed main() {
  cin >> N;
  vector<Point<int>> ps(N);
  for (int i= 0; i < N; ++i) cin >> ps[i];
- sort(ps.begin(), ps.end(), AngleComp<long long>());
+ sort(ps.begin(), ps.end(), AngleComp<int>());
  for (auto &p: ps) cout << p.x << " " << p.y << "\n";
  return 0;
 }
